o2: Perform "reset" on LCD-panel.
LCD-panel documents state we have to perform panel reset after
panel powerup. Sadly we do not have a dedicated GPIO to perform
that hardware reset. Panel hw reset is performed by common system
reset. So on startup we have to:

    - check if we have VAUX3 (lcd power) running
    - switch it ON and reset everything if we don't
    - if we do -- then we have powered and reset the LCD-panel -- proceed with bootup.

That's a way to reset this panel.

HW fix for this is TBDone.