Examine MLO's bootdev markers. Don't reset if MMC.
We have to system reset o2.0 board due to missing dedicated
LCD reset. But after reset the first thing ROMCode does is
sticking to GPMC-CS0 (NAND) as a boot device. When NAND is empty
the system hangs.

So having been started from MMC we want to have a chance to
get to U-Boot prompt to program NAND. So don't reset when MMC,
get to U-boot prompt.